heroku init

+++ server/models/genre.js
@@ -1,7 +1,7 @@
 'use strict';
 module.exports = (sequelize, DataTypes) => {
   const Genre = sequelize.define('Genre', {
-    genreid: DataTypes.BIGINT,
+    genreid: DataTypes.INTEGER,
     name: DataTypes.STRING
   }, {});
   Genre.associate = function(models) {
