+++ server/models/genres.js
@@ -1,9 +1,9 @@
 'use strict';
 module.exports = (sequelize, DataTypes) => {
   const Genres = sequelize.define('Genres', {
-    genresID: DataTypes.BIGINT,
-    mediaID: DataTypes.BIGINT,
-    genreID: DataTypes.BIGINT
+    genresID: DataTypes.INTEGER,
+    mediaID: DataTypes.INTEGER,
+    genreID: DataTypes.INTEGER
   }, {});
   Genres.associate = function(models) {
     // associations can be defined here
