+++ server/models/media.js
@@ -1,8 +1,8 @@
 'use strict';
 module.exports = (sequelize, DataTypes) => {
   const Media = sequelize.define('Media', {
-    mediaID: DataTypes.BIGINT,
-    typeid: DataTypes.BIGINT,
+    mediaID: DataTypes.INTEGER,
+    typeid: DataTypes.INTEGER,
     title: DataTypes.STRING,
     mediaDBID: DataTypes.STRING,
     mediaDBName: DataTypes.STRING
