+++ server/models/reactions.js
@@ -1,11 +1,11 @@
 'use strict';
 module.exports = (sequelize, DataTypes) => {
   const Reactions = sequelize.define('Reactions', {
-    reactionsID: DataTypes.BIGINT,
+    reactionsID: DataTypes.INTEGER,
     table: DataTypes.STRING,
-    tableID: DataTypes.BIGINT,
-    userID: DataTypes.BIGINT,
-    reactionID: DataTypes.BIGINT
+    tableID: DataTypes.INTEGER,
+    userID: DataTypes.INTEGER,
+    reactionID: DataTypes.INTEGER
   }, {});
   Reactions.associate = function(models) {
     // associations can be defined here
