+++ server/models/service.js
@@ -2,8 +2,8 @@
 module.exports = (sequelize, DataTypes) => {
   const Service = sequelize.define('Service', {
     name: DataTypes.STRING,
-    typeID: DataTypes.BIGINT,
-    serviceID: DataTypes.BIGINT
+    typeID: DataTypes.INTEGER,
+    serviceID: DataTypes.INTEGER
   }, {});
   Service.associate = function(models) {
     // associations can be defined here
