+++ server/models/subscription.js
@@ -1,9 +1,9 @@
 'use strict';
 module.exports = (sequelize, DataTypes) => {
   const Subscription = sequelize.define('Subscription', {
-    subscriptionID: DataTypes.BIGINT,
-    userID: DataTypes.BIGINT,
-    serviceID: DataTypes.BIGINT
+    subscriptionID: DataTypes.INTEGER,
+    userID: DataTypes.INTEGER,
+    serviceID: DataTypes.INTEGER
   }, {});
   Subscription.associate = function(models) {
     // associations can be defined here
